put back useEffect to instances component
